new license system

+++ lib/rcs-db/rest/agent.rb
@@ -160,34 +160,38 @@ class AgentController < RESTController
       agent = Item.any_in(_id: @session[:accessible]).find(@params['_id'])
       old_target = agent.get_parent
 
-      evidences = Evidence.collection_class(old_target[:_id]).where(:item => agent[:_id])
+      # factories don't have evidence to be moved
+      if agent._kind == 'agent'
 
-      trace :info, "Moving #{evidences.count} evidence for agent #{agent.name} to target #{target.name}"
+        evidences = Evidence.collection_class(old_target[:_id]).where(:item => agent[:_id])
 
-      # copy the new evidence
-      evidences.each do |e|
-        # deep copy the evidence from one collection to the other
-        ne = Evidence.dynamic_new(target[:_id])
-        Evidence.deep_copy(e, ne)
+        trace :info, "Moving #{evidences.count} evidence for agent #{agent.name} to target #{target.name}"
 
-        # move the binary content
-        if e.data['_grid']
-          bin = GridFS.get(e.data['_grid'], old_target[:_id].to_s)
-          ne.data['_grid'] = GridFS.put(bin, {filename: agent[:_id].to_s}, target[:_id].to_s) unless bin.nil?
+        # copy the new evidence
+        evidences.each do |e|
+          # deep copy the evidence from one collection to the other
+          ne = Evidence.dynamic_new(target[:_id])
+          Evidence.deep_copy(e, ne)
+
+          # move the binary content
+          if e.data['_grid']
+            bin = GridFS.get(e.data['_grid'], old_target[:_id].to_s)
+            ne.data['_grid'] = GridFS.put(bin, {filename: agent[:_id].to_s}, target[:_id].to_s) unless bin.nil?
+          end
+
+          ne.save
         end
 
-        ne.save
+        # remove the old evidence
+        Evidence.collection_class(old_target[:_id]).where(:item => agent[:_id]).destroy_all
       end
-
-      # remove the old evidence
-      Evidence.collection_class(old_target[:_id]).where(:item => agent[:_id]).destroy_all
 
       # actually move the target now.
       # we cant before the evidence move otherwise the grid entries won't get deleted
       agent.path = target.path + [target._id]
       agent.save
       
-      trace :info, "Moving finished for agent #{agent.name}"
+      trace :info, "Moving finished for #{agent._kind} #{agent.name}"
 
       Audit.log :actor => @session[:user][:name],
                 :action => "#{agent._kind}.move",
